make callbacks.py a python pakcage
