+++ src/levanter/__init__.py
@@ -1,4 +1,5 @@
 import levanter.analysis as analysis
+import levanter.callbacks as callbacks
 import levanter.checkpoint as checkpoint
 import levanter.config as config
 import levanter.data as data
