+++ src/levanter/callbacks/__init__.py
@@ -1,172 +1,25 @@
-import abc
-import copy
-import logging as pylogging
 import sys
 import threading
 import time
-from abc import ABC
-from concurrent.futures import ThreadPoolExecutor
-from dataclasses import dataclass
-from datetime import timedelta
-from typing import Any, Callable, Generic, Optional, Sequence, TypeVar
+from typing import Callable, Optional
 
 import jax
-import jax.numpy as jnp
-from jax.tree_util import DictKey, FlattenedIndexKey, GetAttrKey, SequenceKey
-from jaxtyping import PyTree
-from tqdm_loggable import tqdm_logging
 from tqdm_loggable.auto import tqdm
 
-import haliax.nn
-from haliax import NamedArray, is_named_array
-from haliax.jax_utils import is_jax_array_like
-
 import levanter.tracker
-from levanter.data import AsyncDataset, DataLoader
-from levanter.schedule import BatchSchedule
-from levanter.tracker.helpers import log_optimizer_hyperparams
-from levanter.tracker.histogram import Histogram
+from levanter.callbacks._core import Callback, JitCallback, M, S, StepInfo
+from levanter.callbacks._metrics import (
+    _tqdm_logging_one_time_setup,
+    log_epoch_progress,
+    log_performance_stats,
+    log_step_info,
+    logger,
+    pbar_logger,
+)
+from levanter.data import DataLoader
 from levanter.tracker.wandb import WandbConfig
-from levanter.trainer_state import InsideJitInfo, TrainerState
-from levanter.utils import flop_utils, jax_utils
-from levanter.utils.jax_utils import barrier_sync, jnp_to_python
+from levanter.utils.jax_utils import barrier_sync
 from levanter.utils.logging import save_xla_dumps_to_wandb
-
-
-logger = pylogging.getLogger(__name__)
-
-M = TypeVar("M")  # Model
-M_con = TypeVar("M_con", bound=PyTree, contravariant=True)
-S = TypeVar("S", bound=TrainerState)
-CBInfo = TypeVar("CBInfo")
-
-
-@dataclass
-class StepInfo(Generic[S]):
-    """
-    Information about a step that was just completed. This includes the trainer state, the loss, and the duration of the
-    step.
-
-    Note that the step is 0-indexed, so if you want the next step, use `next_step`.
-    """
-
-    state: S
-    loss: float
-    step_duration: float
-
-    model = property(lambda self: self.state.model)
-    opt_state = property(lambda self: self.state.opt_state)
-    eval_model = property(lambda self: self.state.eval_model)
-
-    step = property(lambda self: int(self.state.step) - 1)
-    """
-    The step that was just completed. If you want the next step, use `next_step`.
-    """
-
-    next_step = property(lambda self: int(self.state.step))
-
-
-class Callback(ABC, Generic[S]):
-    """
-    A callback that can be called at the end of a step. This is useful for logging, profiling, and other side effects.
-    """
-
-    @abc.abstractmethod
-    def on_step(self, info: StepInfo[S], force: bool = False):
-        ...
-
-
-class LambdaCallback(Callback[S]):
-    def __init__(self, fn: Callable[[StepInfo[S]], Any]):
-        self.fn = fn
-
-    def on_step(self, info: StepInfo[S], force: bool = False):
-        self.fn(info)
-
-
-class JitCallback(ABC, Generic[S, M, CBInfo]):
-    """
-    A callback that gets called in two phases: inside the step (inside jit), and after the step (outside jit).
-    You have access to [levanter.trainer_state.InsideJitInfo][] inside the step, which allows you to track gradients,
-    updates, etc.
-
-    Note that JitCallbacks tend to be quite heavy, and you should try to line up  all jit callbacks to
-    proc on the same steps.
-
-    The basic flow of a JitCallback is as follows:
-    1. `inside_step` is called inside the JIT-compiled function. This is where you can compute gradients, updates,
-       and other information that you want to track. You should return a PyTree of information that you want to log or use later.
-    2. The returned information from `inside_step` is passed to `on_step` after the JIT-compiled function has completed.
-    """
-
-    @abc.abstractmethod
-    def inside_step(self, state: S, inside_info: InsideJitInfo[M]) -> CBInfo:
-        """
-        This function is called inside the JIT-compiled function. You have access to the `inside_info` which contains
-        information about the gradients, updates, and other information that was computed during the step.
-        Args:
-            state:
-            inside_info:
-
-        Returns:
-
-        """
-        ...
-
-    @abc.abstractmethod
-    def on_step(self, step_info: S, cb_info: CBInfo):
-        """
-        This function is called after the JIT-compiled function has completed. You have access to the `step_info`
-        which contains information about the step that just completed, as well as the `cb_info` which is whatever
-        was returned from `inside_step`.
-        """
-        ...
-
-
-def log_epoch_progress(total_tokens_future, tokens_per_example, batch_size, max_epochs: Optional[int] = None):
-    total_tokens = None
-
-    def log_epoch(step_info: StepInfo):
-        nonlocal total_tokens
-        if total_tokens is None:
-            if not total_tokens_future.done():
-                if step_info.step % 1000 == 0:
-                    logger.info("Dataset not finished. Can't compute epochs.")
-                return  # We don't have the total tokens yet, so we can't calculate epoch
-            total_tokens = total_tokens_future.result()
-
-        # Get the total processed tokens from the metrics logged by log_performance_stats
-        processed_tokens = tokens_per_example * batch_size * step_info.step
-
-        # If we're doing multiple epochs, adjust the denominator
-        total_tokens_for_epochs = total_tokens * max_epochs if max_epochs else total_tokens
-        current_epoch = processed_tokens / total_tokens_for_epochs
-
-        levanter.tracker.log({"train/current_epoch": current_epoch}, step=step_info.step)
-
-    return log_epoch
-
-
-def get_total_dataset_tokens(ds: AsyncDataset, seq_length: int):
-    def log_length():
-        # If ds.async_len() is the only option, run it in an event loop inside the thread
-        import asyncio
-
-        async def compute_length():
-            length = await ds.dataset.async_len()
-            return length
-
-        # Run the async function synchronously in this thread
-        length = asyncio.run(compute_length())
-        total_tokens = length * seq_length
-        levanter.tracker.log_summary({"dataset/total_tokens": total_tokens})
-        return total_tokens
-
-    # Create a ThreadPoolExecutor with a single worker thread
-    executor = ThreadPoolExecutor(max_workers=1)
-    # Submit the log_length function to be executed in a separate thread
-    future = executor.submit(log_length)
-    return future
 
 
 def eval_loss_loop(loss_fn, model, dataset, max_batches: Optional[int] = None, name: Optional[str] = None):
@@ -232,17 +85,6 @@ def compute_validation_loss(
     return compute_loss
 
 
-def log_step_info(total_steps: Optional[int]):
-    def log_step_info_inner(step: StepInfo):
-        metrics = {"train/loss": step.loss, "global_step": step.step}
-        if total_steps:
-            metrics["run_progress"] = step.step / total_steps
-        log_optimizer_hyperparams(step.opt_state, step=step.step, prefix="optim")
-        levanter.tracker.log(metrics, step=step.step)
-
-    return log_step_info_inner
-
-
 def wandb_xla_logger(config: WandbConfig):
     import wandb
 
@@ -258,89 +100,6 @@ def wandb_xla_logger(config: WandbConfig):
         return log_xla_to_wandb
     else:
         return lambda x: None
-
-
-def log_performance_stats(
-    tokens_per_example: int,
-    batch_schedule: int | BatchSchedule,
-    flops_per_example: Optional[float] = None,
-    prefix: Optional[str] = "throughput",
-):
-    if isinstance(batch_schedule, int):
-        batch_schedule = BatchSchedule(batch_schedule)
-
-    def wrap_key(key):
-        if prefix:
-            return f"{prefix}/{key}"
-        return key
-
-    device_count = jax.device_count()
-    device = jax.devices()[0]
-
-    flops_per_device = flop_utils.device_hardware_flops(device)
-
-    if flops_per_device is not None:
-        theoretical_flops = flops_per_device * device_count
-
-        levanter.tracker.log_summary(
-            {
-                wrap_key("device_kind"): device.device_kind,
-                wrap_key("theoretical_flops_per_device"): flops_per_device,
-                wrap_key("theoretical_flops"): theoretical_flops,
-            }
-        )
-
-    if flops_per_example is not None:
-        levanter.tracker.log_summary({wrap_key("flops_per_example"): flops_per_example})
-
-    def log_performance_stats(step_info: StepInfo):
-        dict_to_log: dict[str, float | int] = {}
-
-        # log these totals because it's useful for comparing different seqlens, batch sizes, etc
-        # TODO: if we add seqlen schedules this will get even more complex
-        this_batch_size = batch_schedule.batch_size_at_step(step_info.step)
-        total_examples = batch_schedule.global_data_offset_by_step(step_info.step + 1)
-        total_tokens = tokens_per_example * total_examples
-        dict_to_log["total_tokens"] = total_tokens
-
-        if flops_per_example:
-            total_flops = flops_per_example * total_examples
-            dict_to_log["total_gflops"] = total_flops / 1e9
-
-        if step_info.step_duration != 0.0:
-            dict_to_log["examples_per_second"] = float(this_batch_size) / step_info.step_duration
-            dict_to_log["tokens_per_second"] = float(tokens_per_example) / step_info.step_duration * this_batch_size
-            dict_to_log["duration"] = step_info.step_duration
-
-            if flops_per_example is not None:
-                model_flops_instant = flops_per_example / step_info.step_duration * this_batch_size
-                dict_to_log["gflops_per_second"] = model_flops_instant / 1e9
-
-                if flops_per_device is not None:
-                    mfu_instant = model_flops_instant / theoretical_flops * 100.0
-                    dict_to_log["mfu"] = mfu_instant
-
-        dict_to_log = {wrap_key(k): v for k, v in dict_to_log.items()}
-        levanter.tracker.log(dict_to_log, step=step_info.step)
-
-    return log_performance_stats
-
-
-def pbar_logger(iterable=None, desc="train", **tqdm_mkwargs):
-    kwargs = copy.copy(tqdm_mkwargs)
-    if "desc" not in kwargs:
-        kwargs["desc"] = desc
-    if "iterable" not in kwargs:
-        kwargs["iterable"] = iterable
-
-    _tqdm_logging_one_time_setup()
-    pbar = tqdm(**kwargs)
-
-    def update_pbar(step: StepInfo):
-        pbar.update(step.next_step - pbar.n)
-        pbar.set_postfix(loss=jnp_to_python(step.loss))
-
-    return update_pbar
 
 
 def profile(path: str, start_step: int, num_steps: int, create_perfetto_link: bool) -> Callable[[StepInfo], None]:
@@ -385,251 +144,3 @@ def _flush_while_waiting(event):
 
     thread = threading.Thread(target=flush_stdout)
     thread.start()
-
-
-_did_tqdm_logging_one_time_setup = False
-
-
-def _tqdm_logging_one_time_setup():
-    global _did_tqdm_logging_one_time_setup
-    if _did_tqdm_logging_one_time_setup:
-        return
-    _did_tqdm_logging_one_time_setup = True
-    tqdm_logging.tqdm_logging.set_log_rate(timedelta(seconds=60))
-
-
-class GradWatchCallback(JitCallback[S, M, dict[str, float | Histogram]]):
-    """
-    Emulates the behavior of Wandb's PyTorch-only built-in gradient logging (wandb.watch)
-
-    Args:
-        prefix (str): The prefix to use for logging.
-        include_histogram (bool): Whether to include histograms of the gradients.
-        split_scan_layers (bool): Whether to split the scan layers into separate histograms/norms
-    """
-
-    def __init__(
-        self,
-        prefix: str = "grad",
-        include_histogram: bool = True,
-        split_scan_layers: bool = True,
-    ):
-        self.prefix = prefix
-        self.include_histogram = include_histogram
-        self.split_scan_layers = split_scan_layers
-
-    def inside_step(self, state: TrainerState[M], inside_info: InsideJitInfo[M]):
-        return summary_statistics_for_tree(
-            self.prefix, inside_info.grads, self.split_scan_layers, self.include_histogram
-        )
-
-    def on_step(self, step_info: StepInfo[S], cb_info: dict[str, float | Histogram]):
-        levanter.tracker.log(cb_info, step=step_info.step)
-
-
-class ParamWatchCallback(JitCallback[S, M, dict[str, float | Histogram]]):
-    """
-    Emulates the behavior of Wandb's PyTorch-only built-in gradient logging (wandb.watch)
-
-    Args:
-        prefix (str): The prefix to use for logging.
-        include_histogram (bool): Whether to include histograms of the gradients.
-        split_scan_layers (bool): Whether to split the scan layers into separate histograms/norms
-    """
-
-    def __init__(
-        self,
-        prefix: str = "params",
-        include_histogram: bool = True,
-        split_scan_layers: bool = True,
-    ):
-        self.prefix = prefix
-        self.include_histogram = include_histogram
-        self.split_scan_layers = split_scan_layers
-
-    def inside_step(self, state: TrainerState[M], inside_info: InsideJitInfo[M]):
-        return summary_statistics_for_tree(
-            self.prefix, state.trainable_model, self.split_scan_layers, self.include_histogram
-        )
-
-    def on_step(self, step_info: StepInfo[S], cb_info: dict[str, float | Histogram]):
-        levanter.tracker.log(cb_info, step=step_info.step)
-
-
-class OptStateWatchCallback(JitCallback[S, M, dict[str, float | Histogram]]):
-    """
-    Same as [levanter.callbacks.GradWatchCallback][], but for the optimizer state.
-
-    Args:
-        prefix (str): The prefix to use for logging.
-        include_histogram (bool): Whether to include histograms of the gradients.
-        split_scan_layers (bool): Whether to split the scan layers into separate histograms/norms
-    """
-
-    def __init__(
-        self,
-        prefix: str = "opt_state",
-        include_histogram: bool = True,
-        split_scan_layers: bool = True,
-    ):
-        self.prefix = prefix
-        self.include_histogram = include_histogram
-        self.split_scan_layers = split_scan_layers
-
-    def inside_step(self, state: TrainerState[M], inside_info: InsideJitInfo[M]):
-        to_return = {}
-        # return summary_statistics_for_tree(
-        #     self.prefix, state.opt_state, self.split_scan_layers, self.include_histogram
-        # )
-
-        leaves = jax.tree.leaves_with_path(
-            state.opt_state, is_leaf=lambda m: isinstance(m, type(state.model))
-        )  # noqa: E731
-        for path, v in leaves:
-            if not isinstance(v, type(state.model)):
-                continue
-
-            name = self._munge_key_name(path)
-
-            if name:
-                name_to_log = f"{self.prefix}/{name}"
-            else:
-                name_to_log = self.prefix
-
-            this_hists = summary_statistics_for_tree(name_to_log, v, self.split_scan_layers, self.include_histogram)
-
-            to_return.update(this_hists)
-
-        return to_return
-
-    def on_step(self, step_info: StepInfo[S], cb_info: dict[str, float | Histogram]):
-        levanter.tracker.log(cb_info, step=step_info.step)
-
-    # Optimizer states can have weird/arbitrary structures, but the states we care about
-    # are PyTrees with the same class as our model parameters (e.g., NamedArray, jax arrays, etc.)
-    # opt_state/inner_state/1/nu/ --> we want nu
-    def _munge_key_name(self, path: Sequence[Any]) -> str:
-        if not path:
-            return ""
-        path_elem = path[-1]  # Get the last element in the path, which is the actual value
-        match path_elem:
-            case SequenceKey(idx):  # type: ignore
-                out = f"{idx}"
-            case DictKey(key):  # type: ignore
-                out = f"{key}"
-            case GetAttrKey():  # type: ignore
-                out = str(path_elem)
-            case FlattenedIndexKey(idx):  # type: ignore
-                out = f"{idx}"
-            case _:
-                path_elem = str(path_elem)
-                out = f"{path_elem}"
-
-        if out.startswith("."):
-            out = out[1:]  # Remove leading dot if it exists
-
-        return out
-
-
-class UpdatesWatchCallback(JitCallback[S, M, dict[str, float | Histogram]]):
-    """
-    Same as [levanter.callbacks.GradWatchCallback][], but for the updates to the model parameters.
-
-    Args:
-        prefix (str): The prefix to use for logging.
-        include_histogram (bool): Whether to include histograms of the updates.
-        split_scan_layers (bool): Whether to split the scan layers into separate histograms/norms
-    """
-
-    def __init__(
-        self,
-        prefix: str = "updates",
-        include_histogram: bool = True,
-        split_scan_layers: bool = True,
-    ):
-        self.prefix = prefix
-        self.include_histogram = include_histogram
-        self.split_scan_layers = split_scan_layers
-
-    def inside_step(self, state: TrainerState[M], inside_info: InsideJitInfo[M]):
-        return summary_statistics_for_tree(
-            self.prefix, inside_info.updates, self.split_scan_layers, self.include_histogram
-        )
-
-    def on_step(self, step_info: StepInfo[S], cb_info: dict[str, float | Histogram]):
-        levanter.tracker.log(cb_info, step=step_info.step)
-
-
-def summary_statistics_for_tree(
-    prefix: str, tree: M, split_scan_layers: bool, include_histogram: bool
-) -> dict[str, float | Histogram]:
-    """
-    Computes the summary statistics for a tree of (named) arrays.
-
-    This function is designed to allow you to emulate the behavior of Wandb's PyTorch-only built-in gradient logging,
-    but also works for any PyTree. It computes the Froebinius norm of each array,
-    and optionally the histogram as well.
-
-    Args:
-        prefix: The prefix to use for logging.
-        tree: The tree of arrays to compute the summary statistics for.
-        split_scan_layers: Whether to split the scan layers into separate histograms/norms. Recommended.
-        include_histogram: Whether to include histograms of the gradients. This increases overhead significantly.
-
-    Returns:
-
-    """
-    if split_scan_layers:
-        is_leaf = lambda n: isinstance(n, haliax.nn.Stacked) or is_named_array(n)  # noqa: E731
-    else:
-        is_leaf = is_named_array
-
-    def _rec_log_magnitudes(norms, hists, path_prefix, tree):
-        leaf_key_paths = jax_utils.leaf_key_paths(tree, prefix=path_prefix, is_leaf=is_leaf)
-        del path_prefix
-        for key_path, g in zip(
-            jax.tree.leaves(leaf_key_paths, is_leaf=is_leaf),
-            jax.tree.leaves(tree, is_leaf=is_leaf),
-            strict=True,
-        ):
-            if split_scan_layers and isinstance(g, haliax.nn.Stacked):
-                vmapped_norms, vmapped_hists = haliax.vmap(_rec_log_magnitudes, g.Block)({}, {}, "", g.stacked)
-
-                for k, v in vmapped_norms.items():
-                    for i in range(g.Block.size):
-                        norms[f"{key_path}/{i}/{k}"] = v[i]
-
-                for k, v in vmapped_hists.items():
-                    for i in range(g.Block.size):
-                        hists[f"{key_path}/{i}/{k}"] = jax.tree.map(lambda x: x[i] if is_jax_array_like(x) else x, v)
-
-            elif isinstance(g, NamedArray):
-                # TODO: add linalg.norm to Haliax
-                norms[key_path] = jnp.linalg.norm(g.array)
-                if include_histogram:
-                    hist = Histogram.from_named_array(g)
-                    hists[key_path] = hist
-            elif is_jax_array_like(g):
-                norms[key_path] = jnp.linalg.norm(g)
-
-                if include_histogram:
-                    with jax.named_scope(f"histogram({prefix}/{key_path})"):
-                        hist = Histogram.from_array(g)
-                        hists[key_path] = hist
-
-        return norms, hists
-
-    norms_to_log: dict[str, jax.Array] = {}
-    hists_to_log: dict[str, Histogram] = {}
-
-    _rec_log_magnitudes(norms_to_log, hists_to_log, None, tree)
-
-    to_log: dict = {}
-
-    for key, value in norms_to_log.items():
-        to_log[f"{prefix}/norm/{key}"] = value
-
-    for key, value in hists_to_log.items():
-        to_log[f"{prefix}/hist/{key}"] = value
-
-    return to_log
