+++ src/levanter/main/sft_mixture.py
@@ -193,7 +193,8 @@ def train(config: SFTMixtureConfig):
         flops_per_token = config.model.flops_per_token(vocab_size)
         flops_per_example = 3 * flops_per_token * Pos.size if flops_per_token is not None else None
         trainer.add_hook(
-            callbacks.log_performance_stats(Pos.size, trainer.config.train_batch_size, flops_per_example), every=1
+            callbacks.log_performance_stats(Pos.size, trainer.config.train_batch_size, flops_per_example),
+            every=1,
         )
 
         current_step = int(state.step)
